componentized creating a makerspace
