+++ client/src/pages/site-settings/SiteSettingsPage.tsx
@@ -8,14 +8,7 @@ import PrettyModal from "../../common/PrettyModal";
 import { useState } from "react";
 import AddIcon from '@mui/icons-material/Add';
 import { useNavigate } from "react-router-dom";
-
-const CREATE_MAKERSPACE = gql`
-  mutation CreateMakerspace($name: String!) {
-    addZone(name: $name) {
-      id
-    }
-  }
-`;
+import CreateMakerspaceModal from "./CreateMakerspaceModal";
 
 const ARCHIVE_MAKERSPACE = gql`
   mutation ArchiveMakerspace($id: ID!) {
@@ -29,23 +22,9 @@ export default function SiteSettingsPage() {
   const navigate = useNavigate();
 
   const getZonesResult = useQuery(GET_ZONES);
-
-  const [createMakerspace] = useMutation(CREATE_MAKERSPACE, { refetchQueries: ["GetZones"] });
   const [archiveMakerspace] = useMutation(ARCHIVE_MAKERSPACE, { refetchQueries: ["GetZones"] });
 
   const [createMakerspaceModal, setCreateMakerspaceModal] = useState(false);
-  const [createMakerspaceName, setCreateMakerspaceName] = useState("");
-
-
-  function handleCreateMakerspaceModalClose() {
-    setCreateMakerspaceName("");
-    setCreateMakerspaceModal(false);
-  }
-
-  async function handleCreateMakerspace() {
-    await createMakerspace({ variables: { name: createMakerspaceName } });
-    handleCreateMakerspaceModalClose();
-  }
 
   return (
     <Stack padding={"15px"} width={"100%"} spacing={4}>
@@ -89,36 +68,7 @@ export default function SiteSettingsPage() {
             </Grid>
           );
         }} />
-        <PrettyModal open={createMakerspaceModal} onClose={handleCreateMakerspaceModalClose}>
-          <Stack spacing={2}>
-            <Typography variant="h5">Create Makerspace</Typography>
-            <TextField
-              fullWidth
-              required
-              label="Makerspace Name"
-              value={createMakerspaceName}
-              onChange={(e) => setCreateMakerspaceName(e.target.value)}
-            />
-            <Stack direction={"row"} justifyContent={"space-between"}>
-              <Button
-                color="error"
-                variant="outlined"
-                onClick={handleCreateMakerspaceModalClose}
-              >
-                Cancel
-              </Button>
-              <Button
-                color="success"
-                variant="contained"
-                startIcon={<AddIcon />}
-                disabled={createMakerspaceName === ""}
-                onClick={handleCreateMakerspace}
-              >
-                Create
-              </Button>
-            </Stack>
-          </Stack>
-        </PrettyModal>
+        <CreateMakerspaceModal open={createMakerspaceModal} onClose={() => setCreateMakerspaceModal(false)} />
       </Stack>
     </Stack >
   );
